requirements.txt
Django==2.2.5
django-pagedown==0.1.3
django-crispy-forms==1.7.2
django-markdown-deux==1.0.5